Remove WalletKey from Discovery.Shared
