+++ lib/wallet/cardano-wallet.cabal
@@ -244,6 +244,7 @@ library
     Cardano.Wallet.Address.Keys.MintBurn
     Cardano.Wallet.Address.Keys.PersistPrivateKey
     Cardano.Wallet.Address.Keys.SequentialAny
+    Cardano.Wallet.Address.Keys.Shared
     Cardano.Wallet.Address.Keys.WalletKey
     Cardano.Wallet.Address.Keys.WitnessCount
     Cardano.Wallet.Address.Pool
